requirements.txt
# Core ML/DL frameworks
torch>=1.12.0
torchvision>=0.13.0
pytorch-lightning>=1.6.0
transformers>=4.20.0

# Computer Vision
opencv-python>=4.6.0
Pillow>=9.0.0
albumentations>=1.2.0

# Data handling
numpy>=1.21.0
pandas>=1.4.0
scipy>=1.8.0

# Configuration and utilities
PyYAML>=6.0
omegaconf>=2.2.0
hydra-core>=1.2.0

# Visualization and monitoring
matplotlib>=3.5.0
seaborn>=0.11.0
wandb>=0.12.0
tensorboard>=2.9.0

# Object detection specific
pycocotools>=2.0.4
detectron2 @ git+https://github.com/facebookresearch/detectron2.git

# Development and testing
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
isort>=5.10.0
flake8>=4.0.0
pre-commit>=2.19.0

# Jupyter notebook support
jupyter>=1.0.0
ipywidgets>=7.7.0

# Model evaluation
scikit-learn>=1.1.0
scikit-image>=0.19.0

# Data download and processing
tqdm>=4.64.0
requests>=2.28.0